+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "test",
     "build": "rm -rf dist && NODE_ENV=production webpack --config webpack.config.production.js",
-    "dev": "NODE_ENV=development webpack-dev-server --config webpack.config.development.js --hotOnly",
+    "dev": "NODE_ENV=development webpack-dev-server --config webpack.config.development.js",
     "start-server": "yarn run build && http-server ./dist",
     "deploy-firebase": "yarn run build && firebase deploy --project react-pwa-springboard"
   },
@@ -28,6 +28,7 @@
     "html-webpack-plugin": "^3.2.0",
     "http-server": "^0.11.1",
     "path": "^0.12.7",
+    "react-hot-loader": "^4.8.4",
     "style-loader": "^0.23.1",
     "sw-precache-webpack-plugin": "^0.11.5",
     "webpack": "^4.29.6",
