+++ src/app/flux/configureStore.js
@@ -40,11 +40,11 @@ export default function configureStore(initialState = {}, history) {
 
   // Make reducers hot reloadable, see http://mxs.is/googmo
   /* istanbul ignore next */
-  if (module.hot) {
-    module.hot.accept('./createReducer', () => {
-      store.replaceReducer(createReducer(store.injectedReducers));
-    });
-  }
+  // if (module.hot) {
+  //   module.hot.accept('./createReducer', () => {
+  //     store.replaceReducer(createReducer(store.injectedReducers));
+  //   });
+  // }
 
   return store;
 }
